integrated social db with projects
